Update UART PIC script
I updated the UART PIC script to allow the PIC to receive data. I tested this using Putty on my PC and the Intempco DTG PKIT-2 USB converter. I have the script constantly writing "testing " to the serial port. Then the PIC has an interrupt for when data is written to it and it stores the characters in a buffer array.

I tested this and both functions seem to be working perfectly.